Bug fix from Scott Mobey:
Just a note on this.
This was giving me an error when confirming a email change.
The culprit was in _LoginPartial.cshtml:
Hello @((await UserManager.FindByNameAsync(User.Identity.Name)).FirstName)!
I changed that to this to take care of the error:
@{
var user = await UserManager.GetUserAsync(User);
}
Hello @(user.FirstName)!